requirements.txt
certifi @ file:///croot/certifi_1671487769961/work/certifi
cycler==0.11.0
fonttools==4.38.0
kiwisolver==1.4.4
matplotlib==3.5.3
numpy==1.21.6
packaging==23.1
Pillow==9.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
six==1.16.0
typing_extensions==4.7.1
